feat: 첫 로그인 SuccessHandler 인식 불가 오류 수정
기존: n번째 로그인임에도 첫 로그인으로 인식
현재 : CustomOAuth2UserService에서 CustomOAuth2User 생성순간에 dto로부터
자녀 이름 받아오는 방식 -> 첫 로그인 인식 성공

Resolves: #5

+++ src/main/java/jpabasic/pinnolbe/domain/User.java
@@ -26,7 +26,7 @@ public class User {
     //별도로 입력받아야 할 자녀 정보
     private String childName;
     //자녀 나이
-    private Long childAge;
+    private int childAge;
